+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 4.0.0
 
 * Add the `Pager` API for all the synchronization primitives.
+* `Semaphore::acquire_many*` returns `false` if the specified count is greater than the maximum allowed.
 
 3.3.0